+++ R/chameleons.R
@@ -46,7 +46,7 @@
 #' of the flank occupied by the entire flank patch} }
 #' @author David Firth
 #' @source The data were obtained by Dr Devi Stuart-Fox,
-#' http://www.zoology.unimelb.edu.au/research/groups/animal/labs/stuart-fox/index.php,
+#' \url{https://devistuartfox.com/},
 #' and they are reproduced here with her kind permission.
 #' 
 #' These are the same data that were used in