+++ R/flatlizards.R
@@ -74,7 +74,7 @@
 #' and Blomberg, S. P. (2006). Ultraviolet signals ultra-aggression in a
 #' lizard. *Animal Behaviour* **72**, 353--363.
 #' @source The data were collected by Dr Martin Whiting,
-#' http://whitinglab.com/?page_id=3380, and they appear here with his
+#' \url{http://whitinglab.com/people/martin-whiting/}, and they appear here with his
 #' kind permission.
 #' @keywords datasets
 #' @examples