+++ NEWS.md
@@ -1,3 +1,10 @@
+Changes in BradleyTerry2 1.0-9
+==============================
+
+ * fix bug in setting contrasts in internal function `Diff()` that gave warning under R-devel.
+ * update urls (using https where possible).
+ * fix a couple of `if` statements where argument could be > 1.
+ 
 Changes in BradleyTerry2 1.0-8
 ==============================